Lecture-11-VibeTalker-project/pyproject.toml
[project]
name = "lecture-11-vibetalker-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "openai[voice-helpers]>=1.90.0",
    "pyaudio>=0.2.14",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "speechrecognition>=3.14.3",
]
